+++ src/components/author-page/another-plays/another-plays.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
+import { Button } from 'components/ui/button';
+
 import cn from 'classnames';
 import styles from '../author.module.css';
-
-import { Button } from '../../ui/button';
 
 export const AnotherPlays: FC = () => {
   return (
@@ -12,25 +12,25 @@ export const AnotherPlays: FC = () => {
         <li className={cn(styles.block)}>
           <p className={cn(styles.paragraph)}>Камино норте</p>
           <div className={cn(styles.downloadButton)}>
-            <Button size={'s'} iconPlace={'right'} icon={'arrow-down'} label={'Скачать'} border={'none'}/>
+            <Button size='s' iconPlace='right' icon='arrow-down' label='Скачать' border='none'/>
           </div>
         </li>
         <li className={cn(styles.block)}>
           <p className={cn(styles.paragraph)}>Конкретные разговоры пожилых супругов ни о чём</p>
           <div className={cn(styles.downloadButton)}>
-            <Button size={'s'} iconPlace={'right'} icon={'arrow-down'} label={'Скачать'} border={'none'}/>
+            <Button size='s' iconPlace='right' icon='arrow-down' label='Скачать' border='none'/>
           </div>
         </li>
         <li className={cn(styles.block)}>
           <p className={cn(styles.paragraph)}>Камино норте</p>
           <div className={cn(styles.downloadButton)}>
-            <Button size={'s'} iconPlace={'right'} icon={'arrow-down'} label={'Скачать'} border={'none'}/>
+            <Button size='s' iconPlace='right' icon='arrow-down' label='Скачать' border='none'/>
           </div>
         </li>
         <li className={cn(styles.block)}>
           <p className={cn(styles.paragraph)}>Камино норте</p>
           <div className={cn(styles.downloadButton)}>
-            <Button size={'s'} iconPlace={'right'} icon={'arrow-down'} label={'Скачать'} border={'none'}/>
+            <Button size='s' iconPlace='right' icon='arrow-down' label='Скачать' border='none'/>
           </div>
         </li>
       </ul>
