+++ src/components/author-page/another-plays/index.ts
@@ -1,0 +1,1 @@
+export * from './another-plays';
