--- src/components/author-page/anotherPlays/index.ts
@@ -1,1 +1,0 @@
-export * from './anotherPlays';
