+++ src/components/author-page/information/information.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react';
+import { Button } from 'components/ui/button';
+
 import cn from 'classnames';
 import styles from '../author.module.css';
-
-import { Button } from '../../ui/button';
 
 export const AuthorInformation: FC = () => {
   return (
@@ -11,23 +11,23 @@ export const AuthorInformation: FC = () => {
 
       {/* TODO: Нужно применить компонент "Ссылка" по стайлгайду "Ссылка в таблице", ну и переверстать есессна */}
       <div className={cn(styles.blocksInfo)}>
-        <a className={cn(styles.anchor)} href={'#'}>
+        <a className={cn(styles.anchor)} href='#'>
           <p className={cn(styles.anchorHeading)}>Длинное название статьи, в которой упоминается автор, которое не
             помещается на одну строчку</p>
-          <Button className={cn(styles.anchorButton)} size={'s'} iconPlace={'right'} icon={'arrow-right'} label={''} border={'none'}/>
+          <Button className={cn(styles.anchorButton)} size='s' iconPlace='right' icon='arrow-right' label='' border='none'/>
         </a>
-        <a className={cn(styles.anchor)} href={'#'}>
+        <a className={cn(styles.anchor)} href='#'>
           <p className={cn(styles.anchorHeading)}>Название статьи, в которой упоминается автор</p>
-          <Button className={cn(styles.anchorButton)} size={'s'} iconPlace={'right'} icon={'arrow-right'} label={''} border={'none'}/>
+          <Button className={cn(styles.anchorButton)} size='s' iconPlace='right' icon='arrow-right' label='' border='none'/>
         </a>
-        <a className={cn(styles.anchor)} href={'#'}>
+        <a className={cn(styles.anchor)} href='#'>
           <p className={cn(styles.anchorHeading)}>Длинное название статьи, в которой упоминается автор, которое не
             помещается на одну строчку</p>
-          <Button className={cn(styles.anchorButton)} size={'s'} iconPlace={'right'} icon={'arrow-right'} label={''} border={'none'}/>
+          <Button className={cn(styles.anchorButton)} size='s' iconPlace='right' icon='arrow-right' label='' border='none'/>
         </a>
-        <a className={cn(styles.anchor)} href={'#'}>
+        <a className={cn(styles.anchor)} href='#'>
           <p className={cn(styles.anchorHeading)}>Название статьи, в которой упоминается автор</p>
-          <Button className={cn(styles.anchorButton)} size={'s'} iconPlace={'right'} icon={'arrow-right'} label={''} border={'none'}/>
+          <Button className={cn(styles.anchorButton)} size='s' iconPlace='right' icon='arrow-right' label='' border='none'/>
         </a>
       </div>
     </section>
