+++ src/components/author-page/request/request.tsx
@@ -2,15 +2,15 @@ import { FC } from 'react';
 import cn from 'classnames';
 import styles from '../author.module.css';
 
-import { Icon } from '../../ui/icon';
+import { Icon } from 'components/ui/icon';
 
 export const AuthorRequest: FC = () => {
   return (
     <section className={cn(styles.request)}>
       <div className={cn(styles.footnoteInfo)}>
-        <Icon className={cn(styles.asterisk)} glyph={'asterisk'}/>
+        <Icon className={cn(styles.asterisk)} glyph='asterisk'/>
         <p className={cn(styles.footnote)}>Это ваша страница? Если вы хотите внести изменения, пожалуйста, напишите нам на&nbsp;
-          <a href={'#'} className={cn(styles.footnoteAnchor)}>autors@lubimovka.ru</a>
+          <a href='#' className={cn(styles.footnoteAnchor)}>autors@lubimovka.ru</a>
         </p>
       </div>
     </section>
